Move aCallNotificationData to the tests package of the main module, so that the test module does not have to depend on the impl module.
