+++ features/call/test/build.gradle.kts
@@ -20,7 +20,6 @@ dependencies {
     implementation(projects.libraries.core)
 
     api(projects.features.call.api)
-    implementation(projects.features.call.impl)
     implementation(projects.libraries.matrix.api)
     implementation(projects.libraries.matrix.test)
     implementation(projects.tests.testutils)
